hostapd: Store RADIUS Access-Accept in STA Info
Store RADIUS Access-Accept message in the `sta_info` struct so it can
be retrieved later if needed, e.g. to inspect RADIUS attributes. Also,
add two functions to the RADIUS library to enable iterating over the
attributes in a RADIUS message.

This change is required to support the ability of openwrt's ubus
component to provide RADIUS attributes on the `authorized` event and
the `get_clients` call. This change is motivated by a dynamic tunnel
initiator program I am working on, which needs to know the tunnel
parameters provided by the RADIUS server, however it could be useful
for other purposes as well.